requirements.txt
python-dotenv
langchain
langchain-mcp-adapters
langchain-openai
langchain-google-community